feat: use camera sensor in gazebo and visualize in rviz
Add camera in rviz and select topic. I also have a virtual link for the camera, so it's rotated to have z axis pointing forward (which means x goes from left to right, y from top to bottom) so it has the orientation of images in general, and so it's compatible with things like opencv.